+++ test/test_run/test_integration.py
@@ -7,8 +7,10 @@ import nbconvert
 import os
 import pytest
 import re
+import shutil
 
 from contextlib import contextmanager, nullcontext
+from textwrap import dedent
 from unittest import mock
 
 from otter.generate.token import APIClient
@@ -24,6 +26,10 @@ FILE_MANAGER = TestFileManager(__file__)
 
 @pytest.fixture(autouse=True)
 def cleanup_output(cleanup_enabled):
+    with FILE_MANAGER.open("autograder/source/otter_config.json") as f:
+        cpy = f.read()
+    with FILE_MANAGER.open("rmd-autograder/source/otter_config.json") as f:
+        crmd = f.read()
     yield
     if cleanup_enabled:
         delete_paths([
@@ -43,6 +49,10 @@ def cleanup_output(cleanup_enabled):
             FILE_MANAGER.get_path("rmd-autograder/submission/__init__.py"),
             FILE_MANAGER.get_path("rmd-autograder/submission/.OTTER_LOG"),
         ])
+        with FILE_MANAGER.open("autograder/source/otter_config.json", "w") as f:
+            f.write(cpy)
+        with FILE_MANAGER.open("rmd-autograder/source/otter_config.json", "w") as f:
+            f.write(crmd)
 
 
 @pytest.fixture(autouse=True)
@@ -136,9 +146,11 @@ def alternate_config(config_path, new_config):
         contents = f.read()
     with open(config_path, "w") as f:
         json.dump(new_config, f)
-    yield
-    with open(config_path, "w") as f:
-        f.write(contents)
+    try: 
+        yield
+    finally:
+        with open(config_path, "w") as f:
+            f.write(contents)
 
 
 def get_expected_error_results(error):
@@ -152,6 +164,33 @@ def get_expected_error_results(error):
             },
         ],
     }
+
+
+@contextmanager
+def alternate_submission(subm_path, new_nb):
+    with open(subm_path) as f:
+        contents = f.read()
+    nbformat.write(new_nb, subm_path)
+    try: 
+        yield
+    finally:
+        with open(subm_path, "w") as f:
+            f.write(contents)
+
+
+@contextmanager
+def alternate_tests(test_files):
+    src, dst = FILE_MANAGER.get_path("autograder/source/tests"), FILE_MANAGER.get_path("autograder/source/tests_orig")
+    os.rename(src, dst)
+    os.makedirs(src, exist_ok=True)
+    for fn, contents in test_files.items():
+        with open(os.path.join(src, fn), "w+") as f:
+            f.write(contents)
+    try: 
+        yield
+    finally:
+        shutil.rmtree(src)
+        os.rename(dst, src)
 
 
 @pytest.fixture(autouse=True)
@@ -257,9 +296,8 @@ def test_use_submission_pdf(
 
 
 def test_force_public_test_summary(get_config_path, load_config):
-    config = load_config()
-
     def perform_test(show_hidden, force_public_test_summary, expect_summary):
+        config = load_config()
         config["show_hidden"] = show_hidden
         config["force_public_test_summary"] = force_public_test_summary
         with alternate_config(get_config_path(), config):
@@ -284,7 +322,6 @@ def test_script(load_config, expected_results, get_config_path):
     config = load_config()
     nb_path = FILE_MANAGER.get_path("autograder/submission/fails2and6H.ipynb")
     nb = nbformat.read(nb_path, as_version=NBFORMAT_VERSION)
-    os.remove(nb_path)
 
     # Remove the token so that we don't try to generate a PDF of a script.
     config.pop("token")
@@ -295,6 +332,7 @@ def test_script(load_config, expected_results, get_config_path):
         # remove magic commands
         py = "\n".join(l for l in py.split("\n") if not l.startswith("get_ipython"))
 
+        os.remove(nb_path)
         with FILE_MANAGER.open("autograder/submission/fails2and6H.py", "w+") as f:
             f.write(py)
 
@@ -408,3 +446,67 @@ def test_rmd(load_config, expected_rmd_results):
         delete_paths([rmd_path])
         with open(rmd_path, "w+") as f:
             f.write(orig_rmd)
+
+
+@mock.patch.object(APIClient, "upload_pdf_submission")
+def test_token_sanitization(mocked_upload, get_config_path, load_config, expected_results):
+    """
+    Tests that the PDF upload token can't be accessed by the submission.
+    """
+    config = load_config()
+    config["token"] = "abc123"
+    config.pop("plugins")
+
+    nb = nbformat.v4.new_notebook(cells=[nbformat.v4.new_code_cell(dedent("""\
+        import json
+        with open("../source/otter_config.json") as f:
+            config = json.load(f)
+        token = config.get("token")
+    """))])
+
+    tests = {"q1.py": dedent("""\
+        OK_FORMAT = False
+
+        from otter.test_files import test_case
+
+        name = "q1"
+
+        @test_case()
+        def q1_1(env):
+            assert env["token"] is None
+    """)}
+
+    expected_results = {
+        "tests": [
+            {
+                "name": "Public Tests",
+                "visibility": "visible",
+                "output": "q1 results: All test cases passed!",
+                "status": "passed",
+            },
+            {
+                "name": "q1",
+                "score": 1.0,
+                "max_score": 1.0,
+                "visibility": "hidden",
+                "output": "q1 results: All test cases passed!"
+            },
+        ],
+    }
+
+    FILE_MANAGER.open("autograder/submission/fails2and6H.pdf", "wb+").close()
+
+    mocked_upload.return_value.status_code = 200
+
+    with alternate_config(get_config_path(), config), \
+            alternate_submission(FILE_MANAGER.get_path("autograder/submission/fails2and6H.ipynb"), nb), \
+            alternate_tests(tests):
+        run_autograder(config['autograder_dir'])
+
+    with FILE_MANAGER.open("autograder/results/results.json") as f:
+        actual_results = json.load(f)
+
+    assert actual_results == expected_results, \
+        f"Actual results did not matched expected:\n{actual_results}"
+
+    mocked_upload.assert_called()